\name{gdxrrw-package}
\alias{gdxrrw-package}
\alias{gdxrrw}
\docType{package}
\title{
This package contains different methods to transfer data between GDX file and R
}
\description{
This package implements an interface between GAMS and R.  It
  includes functions to transfer data between GDX and R and a function to
  call gams from within R.
}
\details{
\tabular{ll}{
Package: \tab gdxrrw\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-06-22\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
It contains five methods
	1. igdx(...)
	2. rgdx('gdxfile', ...)
	3. wgdx('gdxfile', ...)
	4. gams('model', ...)
	5. gdxInfo('gdxfile')
}
\author{Original coding by Rishabh Jain.  Adopted and packaged by
Steve Dirkse.  Maintainer: <R@gams.com>}
\references{
Please read document at http://www.gams.com/~steve/gdxmrw.html
}
\keyword{ package }
\keyword{ data }
\keyword{ optimize }
\keyword{ interface }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
