\name{gdxrrw}
\docType{data}
\alias{GMS_VARTYPE_UNKNOWN}
\alias{GMS_VARTYPE_BINARY}
\alias{GMS_VARTYPE_INTEGER}
\alias{GMS_VARTYPE_POSITIVE}
\alias{GMS_VARTYPE_NEGATIVE}
\alias{GMS_VARTYPE_FREE}
\alias{GMS_VARTYPE_SOS1}
\alias{GMS_VARTYPE_SOS2}
\alias{GMS_VARTYPE_SEMICONT}
\alias{GMS_VARTYPE_SEMIINT}
\alias{GMS_VARTYPE_MAX}
\alias{GMS_EQUTYPE_E}
\alias{GMS_EQUTYPE_G}
\alias{GMS_EQUTYPE_L}
\alias{GMS_EQUTYPE_N}
\alias{GMS_EQUTYPE_X}
\alias{GMS_EQUTYPE_C}
\alias{GMS_EQUTYPE_MAX}

\title{Constants used in \code{gdxrrw}}
\description{Constants used in \code{gdxrrw} inputs/outputs, e.g. to
  indicate variable or equation type.}
\keyword{datasets}
