\name{gdxrrw-package}
\alias{gdxrrw-package}
\alias{gdxrrw}
\docType{package}
\title{
  This package contains different methods to transfer data between GDX file and R
}
\description{
This package implements an interface between GAMS and R.  It
  includes functions to transfer data between GDX and R and a function to
  call gams from within R.
}
\details{
\tabular{ll}{
Package: \tab gdxrrw\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2011-08-16\cr
License: \tab GNU GPL\cr
LazyLoad: \tab yes\cr
}
It contains five classes methods:
\itemize{
  \item{igdx(...)}
  \item{rgdx("gdxfile", ...) and related functions}
  \item{wgdx("gdxfile", ...) and related functions}
  \item{gams("gmsfile", ...)}
  \item{gdxInfo("gdxfile")}
}
}
\author{Original coding by Rishabh Jain.  Adopted and packaged by
Steve Dirkse.  Maintainer: \email{R@gams.com}}
\keyword{ package }
\keyword{ data }
\keyword{ optimize }
\keyword{ interface }
\examples{
}
