\name{gdxrrw-package}
\alias{gdxrrw-package}
\alias{gdxrrw}
\docType{package}
\title{
This package contain different methods that import/export data between GDX file and R
}
\description{
This package in interface between GAMS and R. It contain
method to read and write data in the GDX file. Using this 
pacakge the user can also run gams model with R.
}
\details{
\tabular{ll}{
Package: \tab gdxrrw\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-06-22\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
It contain mainly four methods
	1. rgdx('gdxfile', ...)
	2. wgdx('gdxfile', ...)
	3. gams('model', ...)
	4. gdxInfo('gdxfile')
}
\author{
Rishabh Jain

Maintainer: Who to complain to <rishabhjain99@gmail.com>
}
\references{
Please read document at http://www.gams.com/~steve/gdxmrw.html
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
